+++ README.md
@@ -1,1 +1,17 @@
-This is a starter template for [Learn Next.js](https://nextjs.org/learn).
+# セットアップ
+
+- クローン
+
+npx create-next-app boardgamefront --use-npm --example "https://github.com/FukaseDaichi/BoardGameFront/tree/develop"
+
+- アップデート（必要だった）
+
+  npx browserslist@latest --update-db
+
+- 追跡
+
+  git remote add origin https://FukaseDaichi@github.com/FukaseDaichi/BoardGameFront.git
+
+- 実行
+
+  npm run dev
