+++ src/const/next.config.ts
@@ -1,0 +1,13 @@
+/** システム定数クラス */
+export namespace SystemConst {
+	/** サーバー */
+	export namespace Server {
+		/** ホスト名 */
+		export const AP_HOST = "http://localhost:8080/";
+		export const CREATE_ROOM = "createroom";
+	}
+
+	export namespace Message {
+		export const MSG_SYSTEMERR = "エラーが発生しました。";
+	}
+}
